// frontend/src/api-generated/models/EventDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type EventDto = {
    id?: number;
    title?: string | null;
    startDate?: string;
    endDate?: string;
    statusId?: number;
    description?: string | null;
};